All functions ported to workspace_utils, commented, and cleaned.
